services/worker/requirements.txt
fastapi
uvicorn
python-dotenv
pika
redis

opentelemetry-api
opentelemetry-sdk
opentelemetry-exporter-otlp
opentelemetry-instrumentation-fastapi